seperate displayShow from class api
